+++ packages/uikit/src/atoms/FormElements/index.tsx
@@ -1,3 +1,2 @@
 export { default as Textfield } from './Textfield';
 export { default as Textarea } from './Textarea';
-export { default as Select } from './Select';
