+++ packages/uikit/src/index.tsx
@@ -11,7 +11,7 @@ export { default as Icons } from './atoms/Icon/registry';
 export { default as Spacer } from './atoms/Spacer';
 export { default as Responsive } from './atoms/Responsive';
 export { default as Pagination } from './molecules/Pagination';
-export { default as SelectComponent } from './molecules/Select';
+export { default as Select } from './molecules/Select';
 export * from './molecules/Select';
 export * from './util';
